+++ joogopay/__init__.py
@@ -37,7 +37,7 @@ from .types import (
     USDRate,
 )
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"
 
 __all__ = [
     "Client",